Move LockedResolver to its own file, remove defer
